begin another experiment
in bge, I think we only deal with 1 Mesh at a time.

maybe this would simplify the code if we'd replace RAS_MeshObject with
RAS_GameObject and if we put all the code I added in KX_GameObject for
eevee render in this class RAS_GameObject.

This class would contain the current mesh used.

in the API we could maybe skip own.meshes[0].materials[0] and have
own.materials[0] directly instead... I don't really know if this is
feasible and it would need some work but it would be very handy to have
all the code based on GameObject

We could maybe have:

- 1 KX_GameObject/1 RAS_GameObject
- each KX_GameObject would have its own list of KX_BlenderMaterial and each RAS_GameObject would have  its own list of RAS_BlenderMaterial. Each material would have a pointer to its parent GameObject (avoid to share materials, create a new KX_BlenderMaterial per GameObj...)

I don't really know if this is a good organisation but as anyway eevee's
rendering is used directly, for internal bge organization, we maybe
don't necessarly need to mimic blender complexity.

But this would be a big refactor in bge sources and I don't know if I'll
work on it. I just submit the idea, without knowing if it is really
suitable.